+++ source/unstructured.go
@@ -158,7 +158,7 @@ func (us *unstructuredSource) endpointsFromInformer(informer kubeinformers.Gener
 	// Get objects that match the indexer filter (annotation and label selectors)
 	indexKeys := informer.Informer().GetIndexer().ListIndexFuncValues(informers.IndexWithSelectors)
 	if len(indexKeys) == 0 {
-	    return nil, nil
+		return nil, nil
 	}
 	for _, key := range indexKeys {
 		obj, err := informers.GetByKey[*unstructured.Unstructured](informer.Informer().GetIndexer(), key)
